feat: add lazygit functionality to nvim
With the help of neovim-remote, and the correct configuration of toggleterm and lazygit, we can integrate lazygit with neovim.

The configuration all revolves around us being able to use the 'e' key in lazygit to edit files. We would normally end up with a nested terminal session, but by using nvr we can use our local nvim instance instead.